requirements.txt
Flask
flask-restx
Flask-SQLAlchemy
werkzeug
click
WTForms
coverage
itsdangerous
Jinja2
requests
werkzeug
Flask-WTF
sqlalchemy
wheel
mysqlclient
dbutils
pandas
pymysql
mysql-connector-python